AutoMapper && DTOs

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,13 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
+using ProEventos.Application.Dtos;
 using ProEventos.Application.Contratos;
 using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 
 namespace ProEventos.API.Controllers
 {
@@ -28,7 +25,7 @@ namespace ProEventos.API.Controllers
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
                 if (eventos == null) 
-                    return NotFound("Nenhum evento encontrado.");
+                    return NoContent();
                 
                 return Ok(eventos);
             }
@@ -46,7 +43,7 @@ namespace ProEventos.API.Controllers
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
                 if (evento == null) 
-                    return NotFound("Eventos por id não encontrado.");
+                    return NoContent();
                 
                 return Ok(evento);
             }
@@ -64,7 +61,7 @@ namespace ProEventos.API.Controllers
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
                 if (evento == null) 
-                    return NotFound("Evento por tema não encontrado.");
+                    return NoContent();
                 
                 return Ok(evento);
             }
@@ -76,13 +73,13 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
                 if (evento == null) 
-                    return BadRequest("Erro ao tentar adicionar evento.");
+                    return NoContent();
                 
                 return Ok(evento);
             }
@@ -94,13 +91,13 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
                 if (evento == null) 
-                    return BadRequest("Erro ao tentar atualizar evento.");
+                    return NoContent();
                 
                 return Ok(evento);
             }
@@ -116,8 +113,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+                
+
                 return await _eventoService.DeleteEvento(id) ?               
-                    Ok("Deletado") : BadRequest("Evento não deletado");                    
+                    Ok("Deletado") : throw new Exception("Erro ao excluir evento.");                    
                 
             }
             catch (Exception ex)
